warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@11200 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/TSButton.cpp
@@ -292,7 +292,7 @@ nonflat_themed:
 		}
 
 		SIZE sz;
-		GetTextExtentPoint32(hdcMem, szText, mir_tstrlen(szText), &sz);
+		GetTextExtentPoint32(hdcMem, szText, (int)mir_tstrlen(szText), &sz);
 		if (ctl->cHot) {
 			SIZE szHot;
 			GetTextExtentPoint32A(hdcMem, "&", 1, &szHot);
